fixes https://github.com/wso2/product-apim/issues/10672
