+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Listing/components/ImageGenerator/APICards/DocThumbClassic.jsx
@@ -15,208 +15,115 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
+import React, { useState } from 'react';
+// import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
-import Typography from '@material-ui/core/Typography';
-import { FormattedMessage, injectIntl } from 'react-intl';
-import green from '@material-ui/core/colors/green';
+// import Typography from '@material-ui/core/Typography';
+import Grid from '@material-ui/core/Grid';
+import Box from '@material-ui/core/Box';
+// import { FormattedMessage, injectIntl } from 'react-intl';
+import TextFieldsIcon from '@material-ui/icons/TextFields';
+import LetterGenerator from 'AppComponents/Apis/Listing/components/ImageGenerator/LetterGenerator';
+import { makeStyles } from '@material-ui/core/styles';
+import DescriptionTwoToneIcon from '@material-ui/icons/DescriptionTwoTone';
+import LinkIcon from '@material-ui/icons/Link';
+import PictureAsPdfIcon from '@material-ui/icons/PictureAsPdf';
 
-import BaseThumbnail from '../BaseThumbnail';
-
-const styles = (theme) => ({
+const useStyles = makeStyles((theme) => ({
     card: {
         margin: theme.spacing(3 / 2),
-        maxWidth: theme.custom.thumbnail.width,
+        maxWidth: theme.spacing(32),
         transition: 'box-shadow 0.3s ease-in-out',
     },
-    providerText: {
-        textTransform: 'capitalize',
-    },
-    docDetails: { padding: theme.spacing(1) },
-    docActions: { justifyContent: 'space-between', padding: `0px 0px ${theme.spacing(1)}px 0px` },
-    deleteProgress: {
-        color: green[200],
-        position: 'absolute',
-        marginLeft: '200px',
-    },
     thumbHeader: {
-        width: '90%',
-        whiteSpace: 'nowrap',
-        color: theme.palette.text.secondary,
-        overflow: 'hidden',
-        textOverflow: 'ellipsis',
-        cursor: 'pointer',
-        margin: 0,
-        'padding-left': '5px',
-    },
-    imageWrapper: {
-        color: theme.palette.text.secondary,
-        backgroundColor: theme.palette.background.paper,
-        width: theme.custom.thumbnail.width + theme.spacing(1),
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    thumbContent: {
-        width: theme.custom.thumbnail.width - theme.spacing(1),
-        backgroundColor: theme.palette.background.paper,
-        padding: theme.spacing(1),
-    },
-    thumbLeft: {
-        alignSelf: 'flex-start',
-        flex: 1,
-        'padding-left': '5px',
-        'padding-right': '65px',
-    },
-    thumbRight: {
-        alignSelf: 'flex-end',
-    },
-    thumbInfo: {
-        display: 'flex',
-    },
-    contextBox: {
-        width: '110px',
+        maxWidth: theme.spacing(16),
         whiteSpace: 'nowrap',
         overflow: 'hidden',
         textOverflow: 'ellipsis',
-        cursor: 'pointer',
-        margin: 0,
-        display: 'inline-block',
-        lineHeight: '1em',
-        'padding-top': 5,
-        'padding-right': 5,
-        'padding-bottom': 1.5,
-        textAlign: 'left',
     },
-    imageOverlap: {
-        position: 'absolute',
-        bottom: 1,
-        backgroundColor: theme.custom.thumbnail.contentBackgroundColor,
-    },
-    row: {
-        display: 'inline-block',
-    },
-    textWrapper: {
-        color: theme.palette.text.secondary,
-        textDecoration: 'none',
-    },
-    thumbBy: {
-        'padding-left': '5px',
-    },
-    thumbRightBy: {
-        'margin-right': '5px',
-    },
-});
+}));
 
-/**
- *
- * Render Doc Card component in Search listing card view,containing essential Doc information like doc name ,
- * source type ect
- * @class DocThumb
- * @extends {Component}
- */
-class DocThumb extends Component {
-    /**
-     *Creates an instance of DocThumb.
-     * @param {*} props
-     * @memberof DocThumb
-     */
-    constructor(props) {
-        super(props);
-        this.state = { isHover: false };
-        this.toggleMouseOver = this.toggleMouseOver.bind(this);
+
+const DocThumb = (props) => {
+    const { doc } = props;
+    const [isHover, setIsHover] = useState(false);
+    const toggleMouseOver = () => setIsHover(!isHover);
+    const classes = useStyles();
+    let thumbIcon;
+    let PrefixIcon = TextFieldsIcon;
+    if (doc.sourceType === 'FILE') {
+        PrefixIcon = PictureAsPdfIcon;
+        thumbIcon = DescriptionTwoToneIcon;
+    } else if (doc.sourceType === 'URL') {
+        PrefixIcon = LinkIcon;
     }
-
-    /**
-     * Toggle mouse Hover state to set the card `raised` property
-     *
-     * @param {React.SyntheticEvent} event mouseover and mouseout
-     * @memberof DocThumb
-     */
-    toggleMouseOver(event) {
-        this.setState({ isHover: event.type === 'mouseover' });
-    }
-
-    /**
-     * @inheritdoc
-     * @returns {React.Component} @inheritdoc
-     * @memberof DocThumb
-     */
-    render() {
-        const { classes, doc } = this.props;
-        const { isHover } = this.state;
-        return (
-            <Card
-                onMouseOver={this.toggleMouseOver}
-                onFocus={this.toggleMouseOver}
-                onMouseOut={this.toggleMouseOver}
-                onBlur={this.toggleMouseOver}
-                raised={isHover}
-                className={classes.card}
-            >
-                <CardMedia src='None' component={BaseThumbnail} height={140} title='Thumbnail' api={doc} />
-                <CardContent className={classes.apiDetails}>
-                    <div className={classes.textWrapper}>
-                        <Link to='/'>
-                            <Typography gutterBottom variant='h4' className={classes.thumbHeader} title={doc.name}>
+    return (
+        <Card
+            onMouseOver={toggleMouseOver}
+            onFocus={toggleMouseOver}
+            onMouseOut={toggleMouseOver}
+            onBlur={toggleMouseOver}
+            elevation={isHover ? 4 : 1}
+            className={classes.card}
+        >
+            <CardMedia
+                width={200}
+                component={LetterGenerator}
+                height={140}
+                title='Thumbnail'
+                artifact={{ name: 'Doc' }}
+                charLength={3}
+                ThumbIcon={thumbIcon}
+                bgColor={false}
+            />
+            <CardContent>
+                <Grid
+                    container
+                    direction='column'
+                    justify='space-evenly'
+                    alignItems='flex-start'
+                >
+                    <Grid item>
+                        <Box display='flex' alignItems='center' flexDirection='row' fontFamily='fontFamily'>
+                            <Box display='flex'>
+                                <PrefixIcon color='primary' />
+                            </Box>
+                            <Box
+                                className={classes.thumbHeader}
+                                color='text.primary'
+                                fontSize='h4.fontSize'
+                                ml={1}
+                            >
                                 {doc.name}
-                            </Typography>
-                        </Link>
-                    </div>
-                    <div className={classes.row}>
-                        <Typography variant='caption' gutterBottom align='left' className={classes.thumbBy}>
-                            <FormattedMessage
-                                id='Apis.Listing.components.ImageGenerator.DocThumb.sourceType'
-                                defaultMessage='Source Type:'
-                            />
-                            {doc.sourceType}
-                        </Typography>
-                    </div>
-                    <div className={classes.thumbInfo}>
-                        <div className={classes.row}>
-                            <div className={classes.thumbLeft}>
-                                <Typography variant='subtitle1'>{doc.apiName}</Typography>
-                            </div>
+                            </Box>
+                        </Box>
+                    </Grid>
+                    <Grid item>
+                        <Box mt={3} fontFamily='fontFamily'>
+                            <Box color='primary.main'>
+                                {doc.associatedType}
+                            </Box>
+                            <Box color='text.primary' fontSize='h6.fontSize'>
+                                {doc.apiName}
+                            </Box>
+                            <Box color='text.secondary' fontSize='body1.fontSize'>
+                                Version:
+                                {' '}
+                                {doc.apiVersion}
+                            </Box>
+                        </Box>
+                    </Grid>
 
-                            <div className={classes.thumbLeft}>
-                                <Typography variant='caption' gutterBottom align='left'>
-                                    <FormattedMessage
-                                        id='Apis.Listing.components.ImageGenerator.DocThumb.apiName'
-                                        defaultMessage='API Name'
-                                    />
-                                </Typography>
-                            </div>
-                        </div>
-                        <div className={classes.row}>
-                            <div className={classes.thumbRight}>
-                                <Typography variant='subtitle1' align='right' className={classes.contextBox}>
-                                    {doc.apiVersion}
-                                </Typography>
-                            </div>
-
-                            <div className={classes.thumbRight}>
-                                <Typography variant='caption' gutterBottom align='right' className={classes.context}>
-                                    <FormattedMessage
-                                        id='Apis.Listing.components.ImageGenerator.DocThumb.apiVersion'
-                                        defaultMessage='API Version'
-                                    />
-                                </Typography>
-                            </div>
-                        </div>
-                    </div>
-                </CardContent>
-            </Card>
-        );
-    }
-}
+                </Grid>
+            </CardContent>
+        </Card>
+    );
+};
 
 DocThumb.propTypes = {
-    classes: PropTypes.shape({}).isRequired,
     doc: PropTypes.shape({
         id: PropTypes.string,
         name: PropTypes.string,
@@ -225,5 +132,4 @@ DocThumb.propTypes = {
         apiVersion: PropTypes.string.isRequired,
     }).isRequired,
 };
-
-export default injectIntl(withStyles(styles)(DocThumb));
+export default DocThumb;
