+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Apis/Listing/components/ImageGenerator/BaseThumbnail.jsx
@@ -160,7 +160,7 @@ const BaseThumbnail = (props) => {
         <LetterGenerator
             width={width}
             height={height}
-            api={api}
+            artifact={api}
         />
     );
     // If configured the thumbnail variant as `image` or migrated from old thumbnail
